refactor(core): rename parsers to framework-parsers
