+++ .changeset/rename-parser-folder.md
@@ -1,0 +1,6 @@
+---
+'@lapidist/design-lint': patch
+---
+
+rename parser folder to framework-parsers
+
