+++ src/core/framework-parsers/svelte-parser.ts
