+++ src/core/framework-parsers/ts-parser.ts
