+++ src/core/framework-parsers/vue-parser.ts
